@@ -1,18 +1,19 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
 	<OutputType>WinExe</OutputType>
-	<TargetFramework>net7.0-windows10.0.19041.0</TargetFramework>
+	<TargetFramework>net8.0-windows10.0.19041.0</TargetFramework>
 	<TargetPlatformMinVersion>10.0.17763.0</TargetPlatformMinVersion>
 	<RootNamespace>Maestro</RootNamespace>
 	<ApplicationIcon>Assets/WindowIcon.ico</ApplicationIcon>
 	<ApplicationManifest>app.manifest</ApplicationManifest>
 	<Platforms>x86;x64;arm64</Platforms>
-	<RuntimeIdentifiers>win10-x86;win10-x64;win10-arm64</RuntimeIdentifiers>
-	<PublishProfile>Properties\PublishProfiles\win10-$(Platform).pubxml</PublishProfile>
+	<RuntimeIdentifiers>win-x86;win-x64;win-arm64</RuntimeIdentifiers>
+	<PublishProfile>Properties\PublishProfiles\win-$(Platform).pubxml</PublishProfile>
 	<ImplicitUsings>enable</ImplicitUsings>
 	<Nullable>enable</Nullable>
 	<UseWinUI>true</UseWinUI>
 	<EnableMsixTooling>true</EnableMsixTooling>
+	<LangVersion>latestmajor</LangVersion>
   </PropertyGroup>
   <ItemGroup>
     <None Remove="Views\MusicbotPage.xaml" />
@@ -34,6 +35,7 @@
 
   <ItemGroup>
 	<ProjectReference Include="..\Maestro.Core\Maestro.Core.csproj" />
+	<ProjectReference Include="..\Maestro.DiscordBot\Maestro.DiscordBot.csproj" />
   </ItemGroup>
   <ItemGroup>
 	<None Update="appsettings.json">
